chore: refactor build and development scripts
- Added new scripts for build, dev, and start commands using runTurbo utility
- Removed outdated scripts that handled framework-specific logic
- Introduced a utility function to streamline command execution with Turbo
