+++ scripts/start.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { runTurbo } from './utils/turbo.js';
+
+runTurbo('start');
